+++ dora/rest_auth/serializers.py
@@ -1,15 +1,7 @@
 from rest_framework import serializers
 
-from dora.services.models import (
-    Bookmark,
-    SavedSearch,
-    SavedSearchFrequency,
-    ServiceCategory,
-    ServiceFee,
-    ServiceKind,
-    ServiceSubCategory,
-)
-from dora.services.serializers import ServiceListSerializer
+from dora.services.models import Bookmark, SavedSearch
+from dora.services.serializers import SavedSearchListSerializer, ServiceListSerializer
 from dora.sirene.models import Establishment
 from dora.structures.models import Structure
 from dora.structures.serializers import StructureListSerializer
@@ -22,73 +14,6 @@ class BookmarkListSerializer(serializers.ModelSerializer):
     class Meta:
         model = Bookmark
         fields = ["service", "creation_date"]
-
-
-class SavedSearchListSerializer(serializers.ModelSerializer):
-    categories = serializers.SlugRelatedField(
-        slug_field="value",
-        queryset=ServiceCategory.objects.all(),
-        many=True,
-        required=False,
-    )
-    categories_display = serializers.SlugRelatedField(
-        source="categories", slug_field="label", many=True, read_only=True
-    )
-
-    subcategories = serializers.SlugRelatedField(
-        slug_field="value",
-        queryset=ServiceSubCategory.objects.all(),
-        many=True,
-        required=False,
-    )
-    subcategories_display = serializers.SlugRelatedField(
-        source="subcategories", slug_field="label", many=True, read_only=True
-    )
-
-    kinds = serializers.SlugRelatedField(
-        slug_field="value",
-        queryset=ServiceKind.objects.all(),
-        many=True,
-        required=False,
-    )
-    kinds_display = serializers.SlugRelatedField(
-        source="kinds", slug_field="label", many=True, read_only=True
-    )
-
-    fees = serializers.SlugRelatedField(
-        slug_field="value",
-        queryset=ServiceFee.objects.all(),
-        many=True,
-        required=False,
-    )
-    fees_display = serializers.SlugRelatedField(
-        source="fees", slug_field="label", many=True, read_only=True
-    )
-
-    frequency = serializers.SlugRelatedField(
-        slug_field="value",
-        queryset=SavedSearchFrequency.objects.all(),
-        many=False,
-        required=True,
-    )
-
-    class Meta:
-        model = SavedSearch
-        fields = [
-            "id",
-            "city_label",
-            "city_code",
-            "categories",
-            "categories_display",
-            "subcategories",
-            "subcategories_display",
-            "kinds",
-            "kinds_display",
-            "fees",
-            "fees_display",
-            "frequency",
-            "creation_date",
-        ]
 
 
 class UserInfoSerializer(serializers.ModelSerializer):
