--- dora/services/templates/email-base.mjml
@@ -1,83 +1,0 @@
-<mjml lang="fr">
-  <mj-head>
-    <mj-attributes>
-      <mj-all font-family="Arial, Helvetica, sans-serif" color="#555" font-size="15px" line-height="24px" padding="0" />
-      <mj-class
-        name="cta"
-        font-size="15px"
-        line-height="24px"
-        background-color="#5B12EB"
-        color="white"
-        border-radius="4px"
-        width="100%"
-      />
-    </mj-attributes>
-
-    <mj-style inline="inline">
-      h1, h2, h3 { font-family: Arial, Helvetica, sans-serif; color: #555; font-size: 15px; line-height: 24px; padding:
-      0; } h1 { color: #000091; font-size: 24px; font-weight: bold; } h2 { font-weight: bold; margin-bottom: 0 } ul, p {
-      margin-top: 0 } a, a:visited { color: #5B12EB; text-decoration-line: underline; }
-    </mj-style>
-
-    <mj-preview> {% block preview %}{% endblock %} </mj-preview>
-  </mj-head>
-
-  <mj-body background-color="#F5F5F5" width="600px">
-    <!-- Logo -->
-    <mj-section>
-      <mj-column padding="32px 40px">
-        <mj-image
-          src="{{ homepage_url }}/emails/logo-dora-email-new.png"
-          height="60px"
-          width="238px"
-          align="left"
-          alt=""
-        />
-      </mj-column>
-    </mj-section>
-
-    <!-- Contenu -->
-    <mj-section padding-bottom="40px">
-      <mj-column padding="44px 40px" border-radius="24px" background-color="white">
-        <!-- Titre -->
-        <mj-text>
-          <h1>{% block title %}{% endblock %}</h1>
-        </mj-text>
-        <mj-spacer height="24px" />
-
-        <!-- Illustration -->
-        <mj-image
-          src="{{ homepage_url }}/emails/illu-saved-search-notification.png"
-          alt=""
-        />
-        <mj-spacer height="24px" />
-
-        <!-- Texte -->
-        <mj-text> {% block content %} {% endblock %} </mj-text>
-        <mj-spacer height="24px" />
-
-
-        {% block signature %}
-        <mj-text>
-          À très bientôt,
-          <br />
-          <strong>L’équipe DORA</strong>
-        </mj-text>
-        {% endblock %}
-
-        <!-- Footer -->
-        {% block footer %}{% endblock %}
-
-        <mj-spacer height="24px" />
-
-        <mj-text>
-          <div style="color: #555555; background-color: #f0f8ff; border-radius: 12px; padding: 16px 24px">
-            {% block support %} En cas de difficulté, n’hésitez pas à
-            <a style="color: #555" href="{{ support_link }}">contacter le support</a>
-            de la plateforme DORA pour obtenir de l’aide. {% endblock %}
-          </div>
-        </mj-text>
-      </mj-column>
-    </mj-section>
-  </mj-body>
-</mjml>
